Duolingo part 2A af

+++ S7/src/main/kotlin/Main.kt
@@ -1,10 +1,9 @@
 fun main(){
     println("Hoeveel woorden wil je oefenen?")
-    val amountOfWords = readLine()!!
+    val amountOfWords = readLine()!!.toInt()
     println("In welke taal wil je oefenen? Tsjechisch of West-Vlaams?")
     val chooseLanguage = readLine()!!
 
     val duolingo = Duolingo(amountOfWords, chooseLanguage)
     duolingo.play()
-
 }
